requirements.txt
streamlit==1.45.0
openai==1.77.0
python-dotenv==1.1.0
PyPDF2==3.0.1
